# src/v3/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Nova Tag — v3",
    "version": "3.0.0",
    "description": "API v3 - Controle de Acesso RFID Nova Residence\n\n**WebSocket** `/v3/ws` (mesma porta): exige `Authorization: Bearer <TAG_SERVICE_TOKEN>` no upgrade, máximo TAG_WS_V3_MAX_CLIENTS conexões (padrão 5), ping a cada 30 s. Mensagens `{ event, timestamp, data, meta: { version } }`; eventos `connection.established`, `gate.state.changed` `{ numeroDispositivo, state, keepOpen }`, `antenna.connection.changed` `{ numeroDispositivo, connected }`, `tag.read` `{ numeroDispositivo, tag, authorized, reason, direction }` (sem dados pessoais)."
  },
...
  "security": [{ "ApiToken": [] }],
  "components": {
    "securitySchemes": {
      "ApiToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token de acesso à API."
      }
    },
    "schemas": {
      "ResponseMeta": {
        "type": "object",
        "description": "Metadado transversal presente em toda resposta, dentro ou fora do envelope de sucesso/erro.",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Instante em que a resposta foi gerada.",
            "example": "2026-09-28T03:44:40.342Z"
          },
          "requestId": {
            "type": "string",
            "nullable": true,
            "description": "Identificador da requisição, para correlação em logs e suporte.",
            "example": "req-1"
          },
          "version": {
            "type": "string",
            "enum": ["v3"],
            "description": "Versão da API que gerou a resposta."
          },
          "pagination": {
            "type": "object",
            "description": "Presente somente em endpoints que retornam uma coleção paginada.",
            "properties": {
              "limit": { "type": "integer", "minimum": 1, "example": 25 },
              "offset": { "type": "integer", "minimum": 0, "example": 0 },
              "total": { "type": "integer", "minimum": 0, "example": 137 }
            },
            "required": ["limit", "offset", "total"]
          }
        },
        "required": ["timestamp", "requestId", "version"]
      },
      "ProblemDetails": {
        "type": "object",
        "description": "Estrutura padrão de erro da API, com tipo, título, status HTTP e detalhe da ocorrência.",
        "properties": {
          "type": {
            "type": "string",
            "description": "URN fixo identificando o tipo do erro (não é uma URL navegável).",
            "example": "urn:nova-tag:validation-error"
          },
          "title": {
            "type": "string",
            "description": "Resumo legível do tipo de erro.",
            "example": "Dados de entrada inválidos"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP correspondente, redundante com o status da resposta.",
            "example": 400
          },
          "detail": {
            "type": "string",
            "nullable": true,
            "description": "Explicação específica desta ocorrência do erro.",
            "example": "Um ou mais campos da requisição são inválidos."
          },
          "instance": {
            "type": "string",
            "nullable": true,
            "description": "Path da requisição que originou o erro.",
            "example": "/v3/api/health"
          },
          "validationErrors": {
            "type": "array",
            "description": "Presente somente quando type = validation-error.",
            "items": {
              "type": "object",
              "properties": {
                "path": { "type": "string", "example": "/name" },
                "message": { "type": "string", "example": "Too small: expected string to have >=3 characters" }
              },
              "required": ["path", "message"]
            }
          }
        },
        "required": ["type", "title", "status", "detail", "instance"]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [false] },
          "error": { "$ref": "#/components/schemas/ProblemDetails" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "required": ["success", "error", "meta"]
      },
...
      "GateStateData": {
        "type": "object",
        "properties": {
          "state": {
            "type": "string",
            "enum": ["closed", "opening", "open", "closing", "unknown"],
            "example": "closed"
          },
          "keepOpen": {
            "type": "boolean",
            "description": "true = portão travado aberto, sem fechamento automático.",
            "example": false
          }
        },
        "required": [
          "state",
          "keepOpen"
        ]
      },
...
      "AccessVerifyData": {
        "type": "object",
        "description": "Dados de identificação retornados pela verificação de acesso.",
        "properties": {
          "PERMITIDO": { "type": "string" },
          "SEQPESSOA": { "oneOf": [{ "type": "string" }, { "type": "integer" }] },
          "SEQCLASSIFICACAO": { "oneOf": [{ "type": "string" }, { "type": "integer" }] },
          "CLASSIFAUTORIZADA": { "type": "string" },
          "AUTORIZACAOLANC": { "type": "string" },
          "TIPO": { "type": "string" },
          "SEQIDACESSO": { "oneOf": [{ "type": "string" }, { "type": "integer" }] },
          "QUADRA": { "type": "string" },
          "LOTE": { "type": "string" },
          "PANICO": { "type": "string" },
          "MIDIA": { "type": "string" },
          "IDENT": { "type": "string" },
          "SEQVEICULO": { "oneOf": [{ "type": "string" }, { "type": "integer" }] },
          "NOME": { "type": "string" },
          "DESCRICAO": { "type": "string" }
        }
      },
      "TagCacheItem": {
        "type": "object",
        "properties": {
          "tag": { "type": "string" },
          "validatedAt": { "type": "string", "format": "date-time" },
          "isValid": { "type": "boolean" },
          "accessId": { "type": "string" },
          "verifyData": { "$ref": "#/components/schemas/AccessVerifyData" }
        },
        "required": ["tag", "validatedAt", "isValid"]
      },
      "CacheStats": {
        "type": "object",
        "properties": {
          "positiveSize": { "type": "integer" },
          "negativeSize": { "type": "integer" },
          "positiveTimeout": { "type": "integer", "description": "Duração do cache positivo, em milissegundos." },
          "negativeTimeout": { "type": "integer", "description": "Duração do cache negativo, em milissegundos." }
        },
        "required": ["positiveSize", "negativeSize", "positiveTimeout", "negativeTimeout"]
      },
      "ListCacheData": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "enum": ["positive", "negative", "all"] },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/TagCacheItem" } },
          "stats": { "$ref": "#/components/schemas/CacheStats" }
        },
        "required": ["type", "items", "stats"]
      },
      "ListCacheResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "enum": [true] },
          "data": { "$ref": "#/components/schemas/ListCacheData" },
          "meta": { "$ref": "#/components/schemas/ResponseMeta" }
        },
        "required": ["success", "data", "meta"]
      },
      "OpenGateBody": {
        "type": "object",
        "additionalProperties": false,
        "description": "Corpo vazio = fecha sozinho em GATE_AUTO_CLOSE_DEFAULT (15 s). autoCloseTime e keepOpen são exclusivos.",
        "properties": {
          "autoCloseTime": {
            "type": "integer",
            "minimum": 1,
            "description": "Segundos até o fechamento automático (máximo GATE_AUTO_CLOSE_MAX, padrão 120).",
            "example": 15
          },
          "keepOpen": {
            "type": "boolean",
            "enum": [
              true
            ],
            "description": "Mantém o portão aberto até um comando de fechamento."
          }
        }
      },
      "ConfirmBody": {
        "type": "object",
        "properties": {
          "confirm": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "confirm"
        ]
      },
      "GateCommandData": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "open",
              "close"
            ]
          },
          "autoCloseSeconds": {
            "type": "integer",
            "nullable": true,
            "example": 15
          },
          "gate": {
            "$ref": "#/components/schemas/GateStateData"
          }
        },
        "required": [
          "action",
          "autoCloseSeconds",
          "gate"
        ]
      },
      "GateCommandResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/GateCommandData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "RestartData": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "shutdownDelayMs": {
            "type": "integer",
            "example": 2000
          }
        },
        "required": [
          "message",
          "shutdownDelayMs"
        ]
      },
      "RestartResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/RestartData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "ClearCacheData": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "positive",
              "negative",
              "all"
            ]
          },
          "stats": {
            "$ref": "#/components/schemas/CacheStats"
          }
        },
        "required": [
          "type",
          "stats"
        ]
      },
      "ClearCacheResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ClearCacheData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      },
      "RemoveCacheItemData": {
        "type": "object",
        "properties": {
          "tag": {
            "type": "string",
            "example": "0123456789"
          }
        },
        "required": [
          "tag"
        ]
      },
      "RemoveCacheItemResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/RemoveCacheItemData"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "success",
          "data",
          "meta"
        ]
      }
    },
    "responses": {
      "NotFound": {
        "description": "Rota inexistente.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "error": {
                "type": "urn:nova-tag:not-found",
                "title": "Recurso não encontrado",
                "status": 404,
                "detail": "Rota GET /v3/api/rota-que-nao-existe não existe.",
                "instance": "/v3/api/rota-que-nao-existe"
              },
              "meta": {
                "timestamp": "2026-09-28T03:24:14.989Z",
                "requestId": "req-2",
                "version": "v3"
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Não autorizado.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "success": false,
              "error": {
                "type": "urn:nova-tag:unauthorized",
                "title": "Autenticação necessária",
                "status": 401,
                "detail": "Não autorizado.",
                "instance": "/v3/api/gate/state"
              },
              "meta": {
                "timestamp": "2026-09-28T03:24:14.989Z",
                "requestId": "req-3",
                "version": "v3"
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Corpo ou parâmetros inválidos.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "Portão não aceitou o comando (em movimento, antena desconectada ou já encerrando). Não repetir automaticamente.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Portão acionado recentemente. Respeite o header Retry-After (segundos).",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer"
            }
          }
        }
      },
      "Unavailable": {
        "description": "TagValidator indisponível.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/v3/api/cache": {
      "get": {
        "tags": ["Cache"],
        "summary": "Lista o cache de TAGs validadas",
        "description": "Retorna as entradas do cache de validação de TAGs (positivo, negativo ou ambos) e estatísticas de tamanho e tempo de expiração.",
        "operationId": "listCache",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "description": "Filtra por tipo de cache. Omitido retorna todos.",
            "schema": { "type": "string", "enum": ["positive", "negative", "all", "whitelist", "blacklist"], "default": "all" }
          }
        ],
        "responses": {
          "200": {
            "description": "Entradas do cache.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ListCacheResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "503": { "description": "Serviço de validação de TAGs indisponível." }
        }
      },
      "delete": {
        "tags": [
          "Cache"
        ],
        "summary": "Limpar cache de TAGs",
        "description": "Força revalidação na API central. Toda limpeza é registrada em log com o ator.",
        "operationId": "clearCache",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "positive",
                "negative",
                "all",
                "whitelist",
                "blacklist"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cache limpo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearCacheResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/v3/api/gate/open": {
      "post": {
        "tags": [
          "Gate"
        ],
        "summary": "Abrir portão",
        "description": "Aciona o motor físico. Cooldown entre acionamentos (GATE_COOLDOWN_MS, padrão 3 s). O ator repassado pela nova-api em x-actor-id/x-actor-role vai para o log.",
        "operationId": "openGate",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenGateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comando enviado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateCommandResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v3/api/gate/close": {
      "post": {
        "tags": [
          "Gate"
        ],
        "summary": "Fechar portão",
        "description": "Fecha o portão imediatamente. Mesmo cooldown da abertura.",
        "operationId": "closeGate",
        "responses": {
          "200": {
            "description": "Comando enviado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateCommandResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          }
        }
      }
    },
    "/v3/api/gate/restart": {
      "post": {
        "tags": [
          "Gate"
        ],
        "summary": "Reiniciar serviço do portão",
        "description": "Encerra o processo para o Docker reiniciar o container. Exige { \"confirm\": true }.",
        "operationId": "restartGate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Encerramento agendado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestartResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/v3/api/cache/{tag}": {
      "delete": {
        "tags": [
          "Cache"
        ],
        "summary": "Remover TAG do cache",
        "operationId": "removeCacheItem",
        "parameters": [
          {
            "name": "tag",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 32
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TAG removida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemoveCacheItemResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    }
  }
}